requirements.txt
# Core requirements for automotive-algo project
numpy==1.26.0
pandas==2.1.1
python-dotenv==1.0.0
supabase==2.5.0
scikit-learn==1.3.1
matplotlib==3.7.1
seaborn==0.12.2
openai==1.4.0
tabulate==0.9.0
flask==2.3.3
flask-cors==4.0.0
jupyter>=1.0.0
plotly>=5.0.0  # Optional for visualization
pgvector==0.2.3
requests==2.31.0
pydantic==2.4.2
websockets==11.0.3
redis==5.0.1

# Production requirements
gunicorn==21.2.0 